Modify frontend UI

+++ database/migrations/2018_08_20_190341_create_meetings_table.php
@@ -23,7 +23,7 @@ class CreateMeetingsTable extends Migration
             $table->dateTime('end_time')->nullable();
             $table->integer('owner_id');
             $table->text('record')->nullable();
-            $table->smallInteger('status')->default(1);
+            $table->tinyInteger('status')->default(1);
             $table->boolean('request_money')->default(true);
             $table->timestamps();
         });
