+++ database/migrations/2018_08_20_190349_create_attendees_table.php
@@ -16,7 +16,7 @@ class CreateAttendeesTable extends Migration
         Schema::create('attendees', function (Blueprint $table) {
             $table->integer('meeting_id');
             $table->integer('user_id');
-            $table->smallInteger('status')->default(1);
+            $table->tinyInteger('status')->default(1);
             $table->string('absent_reason')->nullable();
             
             $table->datetime('arrive_time')->nullable();
